ai-services/app/api/agents.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from typing import List, Optional
import google.generativeai as genai
from app.core.config import settings

router = APIRouter()

class AgentCreate(BaseModel):
    name: str
    goal: str
    tools: List[str] = []
    systemPrompt: Optional[str] = None

class AgentTask(BaseModel):
    agentId: str
    task: str

@router.post("/create")
async def create_agent(agent: AgentCreate):
    # In a real app, save to database
    return {"message": "Agent created successfully", "agent": agent}

@router.post("/execute")
async def execute_task(task: AgentTask):
    if not settings.GOOGLE_API_KEY:
        raise HTTPException(status_code=400, detail="Google API Key not configured")
    
    # Simple Agent logic using Gemini Function Calling or equivalent
    genai.configure(api_key=settings.GOOGLE_API_KEY)
    model = genai.GenerativeModel('gemini-1.5-flash')
    
    # Placeholder for actual agent reasoning loop
    try:
        response = model.generate_content(f"Execute the following task as an autonomous agent: {task.task}")
        return {"result": response.text, "status": "completed"}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
