ai-services/requirements.txt
fastapi
uvicorn[standard]
openai
google-generativeai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
weaviate-client
huggingface_hub
langchain
python-multipart
jose[cryptography]
passlib[bcrypt]
opencv-python-headless
av
